--- controllers/argocd/configmap.go
@@ -1,44 +1,0 @@
-// Copyright 2019 ArgoCD Operator Developers
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package argocd
-
-import (
-	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
-)
-
-// reconcileConfigMaps will ensure that all ArgoCD ConfigMaps are present.
-func (r *ReconcileArgoCD) reconcileConfigMaps(cr *argoproj.ArgoCD, useTLSForRedis bool) error {
-	if err := r.reconcileArgoConfigMap(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileRedisConfiguration(cr, useTLSForRedis); err != nil {
-		return err
-	}
-
-	if err := r.reconcileRBAC(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileSSHKnownHosts(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileTLSCerts(cr); err != nil {
-		return err
-	}
-
-	return r.reconcileGPGKeysConfigMap(cr)
-}
