--- controllers/argocd/secret.go
@@ -1,68 +1,0 @@
-// Copyright 2019 ArgoCD Operator Developers
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// 	http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package argocd
-
-import (
-	argoproj "github.com/argoproj-labs/argocd-operator/api/v1beta1"
-)
-
-// reconcileClusterSecrets will reconcile all Secret resources for the ArgoCD cluster.
-func (r *ReconcileArgoCD) reconcileClusterSecrets(cr *argoproj.ArgoCD) error {
-	if err := r.reconcileClusterMainSecret(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileClusterCASecret(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileClusterTLSSecret(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileClusterPermissionsSecret(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileGrafanaSecret(cr); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-// reconcileGrafanaSecret will ensure that the Grafana Secret is present.
-func (r *ReconcileArgoCD) reconcileGrafanaSecret(cr *argoproj.ArgoCD) error {
-	if !cr.Spec.Grafana.Enabled {
-		return nil // Grafana not enabled, do nothing.
-	}
-
-	log.Info(grafanaDeprecatedWarning)
-
-	return nil
-}
-
-// reconcileSecrets will reconcile all ArgoCD Secret resources.
-func (r *ReconcileArgoCD) reconcileSecrets(cr *argoproj.ArgoCD) error {
-	if err := r.reconcileClusterSecrets(cr); err != nil {
-		return err
-	}
-
-	if err := r.reconcileArgoSecret(cr); err != nil {
-		return err
-	}
-
-	return nil
-}
